backend/requirements.txt
# FastAPI and Server
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Data Processing
pandas==2.1.4
numpy==1.26.2
python-docx==1.1.0
PyPDF2==3.0.1

# ML Dependencies
scikit-learn==1.3.2
joblib==1.3.2

# NLP
spacy==3.7.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
